Update to move the resetTokenExpireTime check from user.service into auth.service
